fix: change the chiabls input
