+++ go-bindings/Makefile
@@ -1,13 +1,25 @@
-BUILD="$(PWD)/.."
+SRC_DIR=$(PWD)/../src
+BUILD_DIR=$(PWD)/../build
+CACHE_DIR=$(PWD)/cache
 
-CGO_CXXFLAGS="-I$(BUILD)/build/_deps/relic-src/include -I$(BUILD)/build/_deps/relic-build/include -I$(BUILD)/src -std=c++14"
-CGO_LDFLAGS="-L$(BUILD)/build/_deps/sodium-build -L$(BUILD)/build/_deps/relic-build/lib -L$(BUILD)/build/src -lbls -lrelic_s -lsodium"
+CGO_LDFLAGS="-L$(CACHE_DIR)/lib"
+CGO_CXXFLAGS="-I$(CACHE_DIR)"
+
 GO="go"
 COVERAGE_OUTPUT ?= coverage.out
 
-.PHONY: default vet test
+.PHONY: default vet test clean
 
-default: vet test
+default: prepare vet test clean
+
+prepare:
+ifeq ("$(wildcard $(CACHE_DIR))", "")
+	@mkdir -p $(CACHE_DIR)/chiabls $(CACHE_DIR)/lib
+	@cp -r $(BUILD_DIR)/_deps/relic-src/include/* $(CACHE_DIR)
+	@cp -r $(BUILD_DIR)/_deps/relic-build/include/* $(CACHE_DIR)
+	@cp -r $(SRC_DIR)/* $(CACHE_DIR)/chiabls
+	@find $(BUILD_DIR) -name "lib*.a" | xargs -I{} cp {} $(CACHE_DIR)/lib
+endif
 
 fmt:  ## Run go fmt to format Go files
 	$(GO) fmt
@@ -27,6 +39,7 @@ vet:  ## Go vet all project code
 help:  ## Show This Help
 	@for line in $$(cat Makefile | grep "##" | grep -v "grep" | sed  "s/:.*##/:/g" | sed "s/\ /!/g"); do verb=$$(echo $$line | cut -d ":" -f 1); desc=$$(echo $$line | cut -d ":" -f 2 | sed "s/!/\ /g"); printf "%-30s--%s\n" "$$verb" "$$desc"; done
 
-clean:  ## Clean up transient (generated) files
-	CGO_CXXFLAGS=$(CGO_CXXFLAGS) CGO_LDFLAGS=$(CGO_LDFLAGS) $(GO) clean
+clean: ## Clean up transient (generated) files
+	$(GO) clean
 	rm -f $(COVERAGE_OUTPUT)
+	rm -rf $(CACHE_DIR)
