+++ go-bindings/elements.cpp
@@ -11,7 +11,7 @@
 // limitations under the License.
 
 #include <vector>
-#include "bls.hpp"
+#include "chiabls/bls.hpp"
 #include "blschia.h"
 #include "error.h"
 #include "elements.h"
