+++ go-bindings/elements.go
@@ -12,6 +12,8 @@
 
 package blschia
 
+// #cgo LDFLAGS: -lbls -lsodium -lrelic_s
+// #cgo CXXFLAGS: -std=c++14
 // #include <stdbool.h>
 // #include <stdlib.h>
 // #include "elements.h"
