+++ go-bindings/privatekey.cpp
@@ -11,7 +11,7 @@
 // limitations under the License.
 
 #include <vector>
-#include "bls.hpp"
+#include "chiabls/bls.hpp"
 #include "privatekey.h"
 #include "blschia.h"
 #include "error.h"
