+++ go-bindings/schemes.cpp
@@ -11,7 +11,7 @@
 // limitations under the License.
 
 #include <vector>
-#include "bls.hpp"
+#include "chiabls/bls.hpp"
 #include "privatekey.h"
 #include "elements.h"
 #include "schemes.h"
